Updating to latest FlutterFlow output.

+++ lib/pages/menu_page/menu_page_widget.dart
@@ -15,7 +15,7 @@ class MenuPageWidget extends StatefulWidget {
   const MenuPageWidget({super.key});
 
   static String routeName = 'MenuPage';
-  static String routePath = '/menuPage';
+  static String routePath = 'menuPage';
 
   @override
   State<MenuPageWidget> createState() => _MenuPageWidgetState();
@@ -370,7 +370,7 @@ class _MenuPageWidgetState extends State<MenuPageWidget> {
                                                   size: 24.0,
                                                 ),
                                                 Text(
-                                                  'Stands',
+                                                  'Cotação',
                                                   style: FlutterFlowTheme.of(
                                                           context)
                                                       .bodyMedium
@@ -516,6 +516,86 @@ class _MenuPageWidgetState extends State<MenuPageWidget> {
                                                   'Row_navigate_to');
 
                                               context.pushNamed(
+                                                  SubscriptionPlansPageWidget
+                                                      .routeName);
+                                            },
+                                            child: Row(
+                                              mainAxisSize: MainAxisSize.max,
+                                              children: [
+                                                Icon(
+                                                  Icons.nature,
+                                                  color: FlutterFlowTheme.of(
+                                                          context)
+                                                      .primaryText,
+                                                  size: 24.0,
+                                                ),
+                                                Text(
+                                                  'Clube dos 100',
+                                                  style: FlutterFlowTheme.of(
+                                                          context)
+                                                      .bodyMedium
+                                                      .override(
+                                                        font: GoogleFonts.inter(
+                                                          fontWeight:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .bodyMedium
+                                                                  .fontWeight,
+                                                          fontStyle:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .bodyMedium
+                                                                  .fontStyle,
+                                                        ),
+                                                        letterSpacing: 0.0,
+                                                        fontWeight:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .bodyMedium
+                                                                .fontWeight,
+                                                        fontStyle:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .bodyMedium
+                                                                .fontStyle,
+                                                      ),
+                                                ),
+                                              ].divide(SizedBox(width: 9.0)),
+                                            ),
+                                          ),
+                                          Divider(
+                                            thickness: 2.0,
+                                            color: FlutterFlowTheme.of(context)
+                                                .alternate,
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsetsDirectional.fromSTEB(
+                                        0.0, 0.0, 0.0, 12.0),
+                                    child: Container(
+                                      width: 100.0,
+                                      height: 50.0,
+                                      decoration: BoxDecoration(),
+                                      child: Column(
+                                        mainAxisSize: MainAxisSize.max,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceEvenly,
+                                        children: [
+                                          InkWell(
+                                            splashColor: Colors.transparent,
+                                            focusColor: Colors.transparent,
+                                            hoverColor: Colors.transparent,
+                                            highlightColor: Colors.transparent,
+                                            onTap: () async {
+                                              logFirebaseEvent(
+                                                  'MENU_PAGE_PAGE_Row_rsda88p1_ON_TAP');
+                                              logFirebaseEvent(
+                                                  'Row_navigate_to');
+
+                                              context.pushNamed(
                                                   SacPageWidget.routeName);
                                             },
                                             child: Row(
@@ -530,6 +610,86 @@ class _MenuPageWidgetState extends State<MenuPageWidget> {
                                                 ),
                                                 Text(
                                                   'Fale Conosco',
+                                                  style: FlutterFlowTheme.of(
+                                                          context)
+                                                      .bodyMedium
+                                                      .override(
+                                                        font: GoogleFonts.inter(
+                                                          fontWeight:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .bodyMedium
+                                                                  .fontWeight,
+                                                          fontStyle:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .bodyMedium
+                                                                  .fontStyle,
+                                                        ),
+                                                        letterSpacing: 0.0,
+                                                        fontWeight:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .bodyMedium
+                                                                .fontWeight,
+                                                        fontStyle:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .bodyMedium
+                                                                .fontStyle,
+                                                      ),
+                                                ),
+                                              ].divide(SizedBox(width: 9.0)),
+                                            ),
+                                          ),
+                                          Divider(
+                                            thickness: 2.0,
+                                            color: FlutterFlowTheme.of(context)
+                                                .alternate,
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsetsDirectional.fromSTEB(
+                                        0.0, 0.0, 0.0, 12.0),
+                                    child: Container(
+                                      width: 100.0,
+                                      height: 50.0,
+                                      decoration: BoxDecoration(),
+                                      child: Column(
+                                        mainAxisSize: MainAxisSize.max,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceEvenly,
+                                        children: [
+                                          InkWell(
+                                            splashColor: Colors.transparent,
+                                            focusColor: Colors.transparent,
+                                            hoverColor: Colors.transparent,
+                                            highlightColor: Colors.transparent,
+                                            onTap: () async {
+                                              logFirebaseEvent(
+                                                  'MENU_PAGE_PAGE_Row_zs3b4nw8_ON_TAP');
+                                              logFirebaseEvent(
+                                                  'Row_navigate_to');
+
+                                              context.pushNamed(
+                                                  TermosPageWidget.routeName);
+                                            },
+                                            child: Row(
+                                              mainAxisSize: MainAxisSize.max,
+                                              children: [
+                                                Icon(
+                                                  Icons
+                                                      .document_scanner_outlined,
+                                                  color: FlutterFlowTheme.of(
+                                                          context)
+                                                      .primaryText,
+                                                  size: 24.0,
+                                                ),
+                                                Text(
+                                                  'Termos de uso/Política de privacidade',
                                                   style: FlutterFlowTheme.of(
                                                           context)
                                                       .bodyMedium
@@ -588,28 +748,67 @@ class _MenuPageWidgetState extends State<MenuPageWidget> {
                                           onTap: () async {
                                             logFirebaseEvent(
                                                 'MENU_PAGE_PAGE_Row_nkra97ug_ON_TAP');
+                                            Function() _navigate = () {};
                                             logFirebaseEvent(
-                                                'Row_update_app_state');
-                                            FFAppState().deleteUser();
-                                            FFAppState().user = UserStruct();
+                                                'Row_alert_dialog');
+                                            var confirmDialogResponse =
+                                                await showDialog<bool>(
+                                                      context: context,
+                                                      builder:
+                                                          (alertDialogContext) {
+                                                        return AlertDialog(
+                                                          title: Text('Opa!'),
+                                                          content: Text(
+                                                              'Tem certeza que quer sair?'),
+                                                          actions: [
+                                                            TextButton(
+                                                              onPressed: () =>
+                                                                  Navigator.pop(
+                                                                      alertDialogContext,
+                                                                      false),
+                                                              child:
+                                                                  Text('Não'),
+                                                            ),
+                                                            TextButton(
+                                                              onPressed: () =>
+                                                                  Navigator.pop(
+                                                                      alertDialogContext,
+                                                                      true),
+                                                              child:
+                                                                  Text('Sim'),
+                                                            ),
+                                                          ],
+                                                        );
+                                                      },
+                                                    ) ??
+                                                    false;
+                                            if (confirmDialogResponse) {
+                                              logFirebaseEvent(
+                                                  'Row_update_app_state');
+                                              FFAppState().deleteUser();
+                                              FFAppState().user = UserStruct();
 
-                                            FFAppState().deleteServiceId();
-                                            FFAppState().serviceId = '';
+                                              FFAppState().deleteServiceId();
+                                              FFAppState().serviceId = '';
 
-                                            FFAppState().deleteLeadId();
-                                            FFAppState().leadId = '';
+                                              FFAppState().deleteLeadId();
+                                              FFAppState().leadId = '';
 
-                                            safeSetState(() {});
-                                            logFirebaseEvent('Row_auth');
-                                            GoRouter.of(context)
-                                                .prepareAuthEvent();
-                                            await authManager.signOut();
-                                            GoRouter.of(context)
-                                                .clearRedirectLocation();
+                                              safeSetState(() {});
+                                              logFirebaseEvent('Row_auth');
+                                              GoRouter.of(context)
+                                                  .prepareAuthEvent();
+                                              await authManager.signOut();
+                                              GoRouter.of(context)
+                                                  .clearRedirectLocation();
 
-                                            context.goNamedAuth(
-                                                SignPageWidget.routeName,
-                                                context.mounted);
+                                              _navigate = () =>
+                                                  context.goNamedAuth(
+                                                      SignPageWidget.routeName,
+                                                      context.mounted);
+                                            }
+
+                                            _navigate();
                                           },
                                           child: Row(
                                             mainAxisSize: MainAxisSize.max,
